refactor: remover polimorfismo das telas e implementar lógica via switch
