--- src/main/java/com/sepulcrum/model/localidade/cemiterio/controller/GerenciadorCemiterio.java
@@ -1,38 +1,0 @@
-package com.sepulcrum.model.localidade.cemiterio.controller;
-
-public class GerenciadorCemiterio {
-    /* 
-    private TelaRegistroCemiterio trc = new TelaRegistroCemiterio();
-
-    public void setCemiterio(Cemiterio c){
-        c.setNome(trc.inicializarVariaveis());
-        System.out.println("Digite o estado: ");
-        c.setEstado(teclado.nextLine());
-        System.out.println("Digite o cidade: ");
-        c.setCidade(teclado.nextLine());
-        System.out.println("Digite a rua: ");
-        c.setRua(teclado.nextLine());
-        System.out.println("Digite o número do lote: ");
-        c.setNumero(teclado.nextInt());
-        System.out.println("Digite o capacidade máxima: ");
-        c.setCapacidadeMax(teclado.nextInt());
-        System.out.println("Digite o telefone: ");
-        c.setTelefone(teclado.nextLine());
-        System.out.println("Digite o cep: ");
-        c.setCep(teclado.nextLine());
-    }
-    
-    public void getCemiterio(Cemiterio c){
-        if(c != null){ 
-            System.out.println("Nome do cemiterio: " + c.getNome());
-            System.out.println("Estado: " + c.getEstado());
-            System.out.println("Cidade: " + c.getCidade());
-            System.out.println("Rua: " + c.getRua());
-            System.out.println("Numero: " + c.getNumero());
-            System.out.println("Capacidade máxima: " + c.getCapacidadeMax());
-            System.out.println("Telefone: " + c.getTelefone());
-            System.out.println("Cep: " + c.getCep());
-        }
-    }
-    */
-}
