--- src/main/java/com/sepulcrum/model/pessoas/parente/controller/GerenciadorParente.java
@@ -1,44 +1,0 @@
-package com.sepulcrum.model.pessoas.parente.controller;
-/* 
-import java.util.Scanner;
-import com.sepulcrum.model.pessoas.parente.database.CRUDParente;
-import com.sepulcrum.model.pessoas.parente.view.TelaGerenciadorParente;
-
-public class GerenciadorParente {
-    private final TelaGerenciadorParente tp = new TelaGerenciadorParente(new Scanner(System.in));
-    private final CRUDParente crudP = new CRUDParente();
-
-    public void executar(){
-        int opc;
-        do{
-            opc = tp.menu();
-
-            switch (opc) {
-                case 1:
-                    crudP.create();
-                    break;
-                case 2:
-                    crudP.read();
-                    break;
-                case 3:
-                    crudP.listar();
-                    break;
-                case 4:
-                    crudP.update();   
-                    break;
-                case 5:
-                    crudP.delete();
-                    break;
-                case 0:
-                    tp.msgRetornando();
-                    break;
-                default:
-                    if(opc != 0){
-                        tp.msgOpcInvalida();
-                    }
-                    break;
-            }
-        }while (opc != 0);
-    }
-}
-*/
